aider: Add DSL method `template` and instance method `templates` to `lib/lucid/view.rb`.
# Aider chat conversation:

USER: `lib/lucid/view.rb`

ASSISTANT: Sure, I can help with that. Could you please provide more details about the changes you want to make in the `lib/lucid/view.rb` file?

USER: This file defines a DSL for building hierachical view components. I want to add another DSL method to define a template for the view. It takes an optional symbol argument to give the template name (if not specified, it is the main template), and a block which gives the template definition. Since this is a class level method, it should store the name and block in a data structure that can be accessed in instances of the defining class.

ASSISTANT: To add a new DSL method for defining a template, we need to modify the `lib/lucid/view.rb` file.

Here are the steps we need to follow:

1. Define a new class method `template` that takes an optional symbol argument for the template name and a block for the template definition.
2. Initialize a class instance variable `@templates` as a hash to store the template names and their corresponding blocks.
3. In the `template` method, check if a template name is provided. If not, use `:main` as the default template name.
4. Store the template name and block in the `@templates` hash.
5. Define a new instance method `templates` to access the `@templates` hash in instances of the defining class.

Here is the code change:

```ruby
lib/lucid/view.rb
<<<<<<< ORIGINAL
      attr_reader :event_handlers
    end

    config do
=======
      attr_reader :event_handlers

      # ===================================================== #
      #    Templates
      # ===================================================== #

      #
      # Defines a template with a name and a block that gives
      # the template definition.
      #
      def template (name = :main, &block)
        @templates       ||= {}
        @templates[name] = block
      end

      #
      # Access the templates hash. Provides a default if none
      # has been defined.
      #
      def templates
        @templates ||= {}
      end
    end

    config do
>>>>>>> UPDATED
```

This change adds a new DSL method `template` for defining a template with a name and a block. It also adds a new instance method `templates` for accessing the templates in instances of the defining class.