aider: lib/lucid/view.rb:
```diff
def render
-  html = template.render
+  html = template.call
  doc  = Nokogiri::HTML(html)
  doc.to_xhtml(indent: 2, indent_text: ' ')
end
```

spec/lucid/view_spec.rb:

No changes.

# Aider chat conversation:

USER: I ran this command:

 rspec spec/lucid/view_spec.rb

And got this output:

.......***....F..

Pending: (Failures listed here are expected and do not affect your suite's status)

  1) Lucid::View.post class reference defines a post action
     # No reason given
     # ./spec/lucid/view_spec.rb:123

  2) Lucid::View.post with a root path includes the root path
     # No reason given
     # ./spec/lucid/view_spec.rb:137

  3) Lucid::View.post inline definition defines a post action
     # No reason given
     # ./spec/lucid/view_spec.rb:153

Failures:

  1) Lucid::View#render renders the main template
     Got 0 failures and 2 other errors:

     1.1) Failure/Error: html = template.render

          NoMethodError:
            undefined method `render' for #<Proc:0x00007ff4fe8b7428>
          # ./lib/lucid/view.rb:302:in `render'
          # ./spec/lucid/view_spec.rb:227:in `block (3 levels) in <module:Lucid>'

     1.2) Failure/Error: page.save_page(filename)

          NameError:
            undefined local variable or method `page' for #<RSpec::ExampleGroups::LucidView::Render:0x00007ff4fe8c5708>
          # ./spec/spec_helper.rb:22:in `block (2 levels) in <top (required)>'

Finished in 0.0562 seconds (files took 2.25 seconds to load)
17 examples, 1 failure, 3 pending

Failed examples:

rspec ./spec/lucid/view_spec.rb:221 # Lucid::View#render renders the main template

ASSISTANT: The error message indicates that there is an issue with the `render` method in the `Lucid::View` class. The `render` method is trying to call `render` on an object of type `Proc`, which does not have a `render` method.

The second error is due to the `page` variable not being defined in the `RSpec::ExampleGroups::LucidView::Render` context.

Unfortunately, I can't propose a solution without seeing the actual code. Could you please add the `lib/lucid/view.rb` and `spec/lucid/view_spec.rb` files to the chat so I can take a closer look?

USER: I added these *read-write* files: lib/lucid/view.rb, spec/lucid/view_spec.rb

ASSISTANT: Thank you for providing the files.

The error is occurring because the `render` method in the `Lucid::View` class is trying to call `render` on a `Proc` object. In Ruby, `Proc` objects don't have a `render` method.

The `template` method in the `Lucid::View` class is returning a `Proc` object, which is then being used in the `render` method.

To fix this, we need to call the `Proc` object in the `render` method, not try to call `render` on it.

Here is the change that needs to be made:

```ruby
lib/lucid/view.rb
<<<<<<< ORIGINAL
    def render
      html = template.render
      doc  = Nokogiri::HTML(html)
      doc.to_xhtml(indent: 2, indent_text: ' ')
    end
=======
    def render
      html = template.call
      doc  = Nokogiri::HTML(html)
      doc.to_xhtml(indent: 2, indent_text: ' ')
    end
>>>>>>> UPDATED
```

The second error is due to the `page` variable not being defined in the `RSpec::ExampleGroups::LucidView::Render` context. However, I don't see any reference to a `page` variable in the provided `spec/lucid/view_spec.rb` file. It's possible that the error is coming from a different part of your test suite, not shown in the provided files. If you could provide more context or files related to this error, I could provide a more accurate solution.