aider: Refactored `View.template` method to create an instance of the new `Template` class instead of storing the block directly.
# Aider chat conversation:

USER: The View.template method is a DSL method that defines a template. At the moment, it's just storing the given block in a Hash. I want to create a new class Template to encapsulate the block. Rather that storing the block directly, instantiate Template, passing in the block as argument.

ASSISTANT: To implement this change, we need to modify the following files:

1. `lib/lucid/view.rb`: This is where the `View.template` method is defined. We will modify this method to create an instance of the new `Template` class instead of storing the block directly.

2. We will also need to create a new file for the `Template` class. Let's call it `lib/lucid/template.rb`.

Let's start with the changes to `lib/lucid/view.rb`:

```ruby
lib/lucid/view.rb
<<<<<<< ORIGINAL
      #
      # Defines a template with a name and a block that gives
      # the template definition.
      #
      def template (name = :main, &block)
        @templates       ||= {}
        @templates[name] = block
      end

      #
      # Access the templates hash. Provides a default if none
      # has been defined.
      #
      def templates
        @templates ||= {}
      end
=======
      #
      # Defines a template with a name and a block that gives
      # the template definition.
      #
      def template (name = :main, &block)
        @templates       ||= {}
        @templates[name] = Template.new(&block)
      end

      #
      # Access the templates hash. Provides a default if none
      # has been defined.
      #
      def templates
        @templates ||= {}
      end
>>>>>>> UPDATED
```

Now, let's create the new `Template` class in `lib/lucid/template.rb`:

```ruby
lib/lucid/template.rb
<<<<<<< ORIGINAL
=======
module Lucid
  class Template
    def initialize(&block)
      @block = block
    end

    def call(*args)
      @block.call(*args)
    end
  end
end
>>>>>>> UPDATED
```

This `Template` class simply wraps the block and provides a `call` method to execute it.