aider: Update template_spec to pass a view instance when creating a new Template.
# Aider chat conversation:

USER: Also update template_spec

ASSISTANT: To update the `template_spec.rb` file, we need to modify the tests to reflect the changes in the `Template` class. Specifically, we need to update the tests to pass a view instance when creating a new `Template`.

Here are the changes:

```ruby
spec/lucid/template_spec.rb
<<<<<<< ORIGINAL
    it "initializes with a block" do
      block = Proc.new { "Hello, World" }
      template = Template.new(&block)
      expect(template).to be_a(Template)
    end

    it "calls the block when call method is invoked" do
      block = Proc.new { "Hello, World" }
      template = Template.new(&block)
      expect(template.call).to eq("Hello, World")
    end
=======
    it "initializes with a view and a block" do
      view = double("View")
      block = Proc.new { "Hello, World" }
      template = Template.new(view, &block)
      expect(template).to be_a(Template)
    end

    it "calls the block when call method is invoked" do
      view = double("View")
      block = Proc.new { "Hello, World" }
      template = Template.new(view, &block)
      expect(template.call).to eq("Hello, World")
    end
>>>>>>> UPDATED
```

This change updates the tests to create a mock view object using `double` and pass it when creating a new `Template`. The tests still check that a `Template` is created and that the `call` method works as expected.